Refactor how strikes and spares are handled
While I did this refactoring I kept the original logic around.  I was having trouble reasoning about the logic flow with the original logic, but I'm not a fan of the new logic since I didn't feel there were good methods to pull out of it yet.  I was tempted to try creating a method like `returnScoreFromNextBowls(Frame frame, int numberOfBowls)`, but saw that getting complex with edge cases and figured I'd done enough with this.